refactor: 카드 정보를 출력할 때 활용하는 확장함수를 util 패키지에서 관리
