+++ src/main/kotlin/blackjack/model/Card.kt
@@ -1,5 +1,7 @@
 package blackjack.model
 
+import blackjack.util.display
+
 class Card(val pattern: Pattern, val number: CardNumber) {
     override fun toString(): String {
         return "${number.display()}${pattern.display()}"
