+++ src/main/kotlin/blackjack/model/CardNumber.kt
@@ -15,21 +15,3 @@ enum class CardNumber(val value: Int) {
     QUEEN(10),
     KING(10);
 }
-
-fun CardNumber.display(): String {
-    return when (this) {
-        CardNumber.ACE -> "A"
-        CardNumber.TWO -> "2"
-        CardNumber.THREE -> "3"
-        CardNumber.FOUR -> "4"
-        CardNumber.FIVE -> "5"
-        CardNumber.SIX -> "6"
-        CardNumber.SEVEN -> "7"
-        CardNumber.EIGHT -> "8"
-        CardNumber.NINE -> "9"
-        CardNumber.TEN -> "10"
-        CardNumber.JACK -> "J"
-        CardNumber.QUEEN -> "Q"
-        CardNumber.KING -> "K"
-    }
-}
